requirements.txt
flask==2.1.2
gunicorn==20.1.0
pytest==7.1.1
pytest-cov==3.0.0
pyYAML==6.0
requests==2.27.1
